Fixed warnings on windows.
git-svn-id: https://svn.opendtect.org/od/trunk@27238 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/Algo/bodyvolumecalc.cc
@@ -55,7 +55,7 @@ bool BodyVolumeCalculator::doWork( od_int64 start, od_int64 stop, int threadid )
     {
 	for ( int idy=0; idy<crlsz-1; idy++ )
 	{
-	    for ( int idz=start; idz<stop; idz++ )
+	    for ( int idz=(int) start; idz<stop; idz++ )
 	    {
 		bool inside[8];
 		inside[0] = arr_.get(idx,idy,idz) <= threshold_;
